+++ Source/Bogus/ValidationResult.cs
@@ -2,18 +2,23 @@
 
 namespace Bogus
 {
-  /// <summary>
-  /// Contains validation results after validation
-  /// </summary>
-  internal class ValidationResult
-  {
     /// <summary>
-    /// True if is valid
+    /// Contains validation results after validation
     /// </summary>
-    internal bool IsValid { get; set; }
-    /// <summary>
-    /// A complete list of missing rules
-    /// </summary>
-    internal List<string> MissingRules { get; } = new List<string>();
-  }
+    internal class ValidationResult
+    {
+        /// <summary>
+        /// True if is valid
+        /// </summary>
+        internal bool IsValid { get; set; }
+        /// <summary>
+        /// A complete list of missing rules
+        /// </summary>
+        internal List<string> MissingRules { get; } = new List<string>();
+
+        /// <summary>
+        /// Extra validation messages to display
+        /// </summary>
+        internal List<string> ExtraMessages { get; } = new List<string>();
+    }
 }